WIP Staticisor and CrtControl panels
